Refactor toast imports and usage across multiple components
