+++ app/contact/page.tsx
@@ -30,7 +30,7 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
-import { toast } from "@/components/ui/use-toast";
+import { toast } from "@/hooks/use-toast";
 import NavBar from "@/components/home/NavBar";
 import { Footer } from "@/components/footer";
