+++ app/dashboard/company/employees/page.tsx
@@ -10,7 +10,7 @@ import {
 import { Button } from "@/components/ui/button"
 import { AppAvatar } from "@/components/ui/avatar"
 import { getImage } from "@/lib/utils"
-import { useToast } from "@/components/ui/use-toast"
+import { useToast } from "@/hooks/use-toast"
 import { Textarea } from "@/components/ui/textarea"
 
 const getImageUrl = (profileImage: any) => {
