+++ app/dashboard/company/jobs/[id]/page.tsx
@@ -12,7 +12,7 @@ import { Input } from "@/components/ui/input";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
 import { JOB_CATEGORIES } from "@/app/dashboard/company/post-job/page";
 import { Textarea } from "@/components/ui/textarea";
-import { toast } from "@/components/ui/use-toast";
+import { toast } from "@/hooks/use-toast";
 import { fetchJobById, updateJob } from "@/lib/api";
 import { X } from "lucide-react";
